requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
httpx
